requirements.txt
flask>=3.0.0
langchain-core>=1.0.1
langchain-google-genai>=3.0.0
langchain-huggingface>=1.0.0
langchain-chroma>=1.0.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.3
sentence-transformers>=2.2.2
python-dotenv>=1.0.1
pypdf>=4.2.0
python-json-logger>=2.0.7
langdetect>=1.0.9
transformers>=4.44.0
openai-whisper>=20231117
rank-bm25>=0.2.2
beautifulsoup4>=4.12.3
python-docx>=1.1.0
reportlab>=4.2.0
pytest>=8.3.0
